Update to .NET5 and refactor CI/CD tooling
