--- .travis.yml
@@ -1,34 +1,0 @@
-language: bash
-services:
-- docker
-jobs:
-  include:
-    - stage: "Test"
-      name: "Run and cover tests"
-      scripts: docker run --rm -w /app --env-file <(env) -v $(pwd)/:/app mcr.microsoft.com/dotnet/core/sdk:3.1 /app/scripts/test/cover.sh
-      deploy:
-        provider: releases
-        prerelease: true
-        api_key:
-          secure: VkjSRVs5ZYSYyvWTpBv5yZnrog9FKPdy6iW5OUtCg6Q6Rs4qi4QNlRv9NXXp5zlYFXE2NPAcQVllQG2D/e0X86y/OZHPgkABzIl4ppE3PN0h42QJY7kC9k/CRRBs/osqyuHwmq56vRdbyXlopkReL+s6B19UqdoYXPBmuXuQtSBdzMIzxno3vzFSxQEGulfVJTAKStvXQpY7cyy8FOcu5vmQ0MParArFibiHGcJrCyu9lATg0Ko32f0m2D8XuONiC9uMgckr2hie415RjccsxusNSOHcaDfxNt5FGAj9YJrG+u0lV7YCSzAYdIJKZShpmGxBRSZZlLUCT5kHZwa57uKMv9tbuyFDE3GYFqn4+kRTjIrsw33ZGMgTDC8prPybvlF7czYT/6aZP2Y/PPLgVb8+1TAOrR5IZg5SA0RSaWVDSv1hU+0ZIo2CqkUXsuzLU1N1/YoHclMgmMFgLzvYlRGbchehPSdPgzm3DoUIwqML1EVBcWIsnpZ5/ktHe/qJtrHNiqAziWPBBrSBMpoNx1ZxKo7dLAJcSuy157CSUH09mu5M+Wc675vBgsHS0ysW1ad0u3TYFkN7R5KxwGmlIaICZqcuNvE2tzvl1csyvHxun8XBTIKAyyLVE6fijZtm0BB1kHn2wq4y3brYaTdjG4OMbfBkoEPTd6EvzoikVz0=
-        file:
-          - "./artifacts/coverage/coverage-report.tgz"
-        on:
-          repo: tsqllint/tsqllint
-          tags: true
-    - stage: "Publish"
-      name: "Generate assemblies and publish artifacts"
-      script: docker run --rm -w /app -v $(pwd)/:/app mcr.microsoft.com/dotnet/core/sdk:3.1 /app/scripts/build/publish.sh
-      deploy:
-        provider: releases
-        prerelease: true
-        api_key:
-          secure: VkjSRVs5ZYSYyvWTpBv5yZnrog9FKPdy6iW5OUtCg6Q6Rs4qi4QNlRv9NXXp5zlYFXE2NPAcQVllQG2D/e0X86y/OZHPgkABzIl4ppE3PN0h42QJY7kC9k/CRRBs/osqyuHwmq56vRdbyXlopkReL+s6B19UqdoYXPBmuXuQtSBdzMIzxno3vzFSxQEGulfVJTAKStvXQpY7cyy8FOcu5vmQ0MParArFibiHGcJrCyu9lATg0Ko32f0m2D8XuONiC9uMgckr2hie415RjccsxusNSOHcaDfxNt5FGAj9YJrG+u0lV7YCSzAYdIJKZShpmGxBRSZZlLUCT5kHZwa57uKMv9tbuyFDE3GYFqn4+kRTjIrsw33ZGMgTDC8prPybvlF7czYT/6aZP2Y/PPLgVb8+1TAOrR5IZg5SA0RSaWVDSv1hU+0ZIo2CqkUXsuzLU1N1/YoHclMgmMFgLzvYlRGbchehPSdPgzm3DoUIwqML1EVBcWIsnpZ5/ktHe/qJtrHNiqAziWPBBrSBMpoNx1ZxKo7dLAJcSuy157CSUH09mu5M+Wc675vBgsHS0ysW1ad0u3TYFkN7R5KxwGmlIaICZqcuNvE2tzvl1csyvHxun8XBTIKAyyLVE6fijZtm0BB1kHn2wq4y3brYaTdjG4OMbfBkoEPTd6EvzoikVz0=
-        file:
-          - "./artifacts/assemblies/linux-x64.tgz"
-          - "./artifacts/assemblies/osx-x64.tgz"
-          - "./artifacts/assemblies/win-x64.tgz"
-          - "./artifacts/assemblies/win-x86.tgz"
-        on:
-          repo: tsqllint/tsqllint
-          tags: true
