+++ scripts/cover.sh
