+++ source/TSQLLint.Tests/FunctionalTests/ConsoleAppTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using System.Runtime.InteropServices;
 using NUnit.Framework;
 using TSQLLint.Tests.Helpers;
 
@@ -16,13 +17,6 @@ namespace TSQLLint.Tests.FunctionalTests
         [SetUp]
         public void Setup()
         {
-            if (Environment.OSVersion.Platform == PlatformID.Win32S ||
-                Environment.OSVersion.Platform == PlatformID.Win32Windows ||
-                Environment.OSVersion.Platform == PlatformID.Win32NT)
-            {
-                Assert.Ignore("Tests ignored on windows due to file path issues");
-            }
-
             testDirectoryPath = TestContext.CurrentContext.WorkDirectory;
         }
 
@@ -60,13 +54,13 @@ namespace TSQLLint.Tests.FunctionalTests
         [TestCase(@"TestFiles/linting-disabled.sql", 0)]
         public void LintingPerformedExitCodeTest(string testFile, int expectedExitCode)
         {
-            // var fileLinted = false;
+             var fileLinted = false;
             void OutputHandler(object sender, DataReceivedEventArgs args)
             {
-                // if (args.Data != null && args.Data.Contains("Linted 1 files"))
-                // {
-                //     fileLinted = true;
-                // }
+                 if (args.Data != null && args.Data.Contains("Linted 1 files"))
+                 {
+                     fileLinted = true;
+                 }
             }
 
             void ErrorHandler(object sender, DataReceivedEventArgs args) { }
@@ -83,7 +77,7 @@ namespace TSQLLint.Tests.FunctionalTests
             var process = ConsoleAppTestHelper.GetProcess($"-c {configFilePath} {path}", OutputHandler, ErrorHandler, ExitHandler);
             ConsoleAppTestHelper.RunApplication(process);
 
-            // Assert.IsTrue(fileLinted);
+            Assert.IsTrue(fileLinted);
         }
 
         [TestCase(@"-l", "Loaded plugin: 'TSQLLint.Tests.UnitTests.PluginHandler.TestPlugin'", 0)]
