+++ source/TSQLLint.Tests/Helpers/TestHelper.cs
@@ -2,8 +2,8 @@ using System.Collections.Generic;
 using System.Linq;
 using NSubstitute;
 using NUnit.Framework;
+using TSQLLint;
 using TSQLLint.Common;
-using TSQLLint.Console;
 using TSQLLint.Infrastructure.Rules.RuleViolations;
 using TSQLLint.Tests.Helpers.ObjectComparers;
