+++ source/TSQLLint.Tests/UnitTests/CommandLineOptions/HandlingStrategies/CreateConfigFileStrategyTests.cs
@@ -12,16 +12,6 @@ namespace TSQLLint.Tests.UnitTests.CommandLineOptions.HandlingStrategies
     [TestFixture]
     public class CreateConfigFileStrategyTests
     {
-        [SetUp]
-        [ExcludeFromCodeCoverage]
-        public void Setup()
-        {
-            if (Environment.OSVersion.Platform == PlatformID.MacOSX || Environment.OSVersion.Platform == PlatformID.Unix)
-            {
-                Assert.Ignore("Tests ignored on osx or linux until https://github.com/tathamoddie/System.IO.Abstractions/issues/252 is resolved");
-            }
-        }
-
         [Test]
         public void HandleCommandLineOptions_ConfigFileExistsNoForceParam_ShouldReportError()
         {
