+++ source/TSQLLint.Tests/UnitTests/PluginHandler/PluginHandlerTests.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.IO.Abstractions;
 using System.IO.Abstractions.TestingHelpers;
 using System.Reflection;
+using System.Runtime.InteropServices;
 using NSubstitute;
 using NUnit.Framework;
 using TSQLLint.Common;
@@ -16,26 +17,18 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
 {
     public class PluginHandlerTests
     {
-        [SetUp]
-        [ExcludeFromCodeCoverage]
-        public void Setup()
-        {
-            if (Environment.OSVersion.Platform == PlatformID.MacOSX || Environment.OSVersion.Platform == PlatformID.Unix)
-            {
-                Assert.Ignore("Tests ignored on osx or linux until https://github.com/tathamoddie/System.IO.Abstractions/issues/252 is resolved");
-            }
-        }
+        private readonly bool isExecutingOnWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
 
         [Test]
         public void LoadPlugins_ShouldLoadPluginsFromPathAndFile()
         {
             // arrange
-            const string filePath1 = @"c:\pluginDirectory\plugin_one.dll";
-            const string filePath2 = @"c:\pluginDirectory\plugin_two.dll";
-            const string filePath3 = @"c:\pluginDirectory\plugin_three.dll";
-            const string filePath4 = @"c:\pluginDirectory\foo.txt";
-            const string filePath5 = @"c:\pluginDirectory\subDirectory\bar.txt";
-            const string filePath6 = @"c:\pluginDirectory\subDirectory\plugin_four.dll";
+            var filePath1 = GetTestFilePath(@"c:\pluginDirectory\plugin_one.dll");
+            var filePath2 = GetTestFilePath(@"c:\pluginDirectory\plugin_two.dll");
+            var filePath3 = GetTestFilePath(@"c:\pluginDirectory\plugin_three.dll");
+            var filePath4 = GetTestFilePath(@"c:\pluginDirectory\foo.txt");
+            var filePath5 = GetTestFilePath(@"c:\pluginDirectory\subDirectory\bar.txt");
+            var filePath6 = GetTestFilePath(@"c:\pluginDirectory\subDirectory\plugin_four.dll");
 
             var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
             {
@@ -61,10 +54,10 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
 
             var assemblyWrapper = new TestAssemblyWrapper(new Dictionary<string, int>
             {
-                { @"c:\pluginDirectory\plugin_one.dll", 0 },
-                { @"c:\pluginDirectory\plugin_two.dll", 1 },
-                { @"c:\pluginDirectory\plugin_three.dll", 2 },
-                { @"c:\pluginDirectory\subDirectory\plugin_four.dll", 3 }
+                { GetTestFilePath(@"c:\pluginDirectory\plugin_one.dll"), 0 },
+                { GetTestFilePath(@"c:\pluginDirectory\plugin_two.dll"), 1 },
+                { GetTestFilePath(@"c:\pluginDirectory\plugin_three.dll"), 2 },
+                { GetTestFilePath(@"c:\pluginDirectory\subDirectory\plugin_four.dll"), 3 }
             });
 
             var reporter = Substitute.For<IReporter>();
@@ -75,7 +68,7 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
             var pluginPaths = new Dictionary<string, string>
             {
                 {
-                    "my-first-plugin", @"c:\pluginDirectory\"
+                    "my-first-plugin", GetTestFilePath(@"c:\pluginDirectory\")
                 },
                 {
                     "my-second-plugin", filePath6
@@ -122,7 +115,7 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
             var pluginPaths = new Dictionary<string, string>
             {
                 {
-                    "my-second-plugin", @"..\plugin_one.dll"
+                    "my-second-plugin", @"..\plugin_two.dll"
                 }
             };
 
@@ -138,10 +131,11 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
         public void LoadPlugins_ShouldLoadPluginsDirectoriesWithRelativePaths()
         {
             // arrange
+            var pluginName = "plugin_one.dll";
             var currentDirectory = Directory.GetParent(Assembly.GetExecutingAssembly().Location);
             var executingLocationParentFolder = currentDirectory.Parent.FullName;
 
-            var filePath1 = Path.Combine(executingLocationParentFolder, "subDirectory", "plugin_one.dll");
+            var filePath1 = Path.Combine(executingLocationParentFolder, "subDirectory", pluginName);
 
             var fileSystem = new MockFileSystem(
                 new Dictionary<string, MockFileData>
@@ -162,7 +156,7 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
             var pluginPaths = new Dictionary<string, string>
             {
                 {
-                    "my-first-plugin", @"..\subDirectory\"
+                    pluginName, @"..\subDirectory"
                 }
             };
 
@@ -178,7 +172,7 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
         public void LoadPlugins_ThrowErrors_When_Same_Type_Is_Loaded_More_Than_Once()
         {
             // arrange
-            const string filePath1 = @"c:\pluginDirectory\plugin_one.dll";
+            var filePath1 = GetTestFilePath(@"c:\pluginDirectory\plugin_one.dll");
             var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
             {
                 {
@@ -194,10 +188,10 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
                     "my-plugin", filePath1
                 },
                 {
-                    "my-plugin-directories", @"c:\pluginDirectory"
+                    "my-plugin-directories", GetTestFilePath(@"c:\pluginDirectory")
                 },
                 {
-                    "my-plugin-invalid-path", @"c:\doesnt-exist"
+                    "my-plugin-invalid-path", GetTestFilePath(@"c:\doesnt-exist")
                 }
             };
 
@@ -221,7 +215,7 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
         public void ActivatePlugins_PluginRuleViolations_ShouldCallReporter()
         {
             // arrange
-            const string filePath1 = @"c:\pluginDirectory\plugin_one.dll";
+            var filePath1 = GetTestFilePath(@"c:\pluginDirectory\plugin_one.dll");
             var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
             {
                 {
@@ -240,7 +234,9 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
 
             var reporter = Substitute.For<IReporter>();
             var textReader = ParsingUtility.CreateTextReaderFromString("\tSELECT * FROM FOO");
-            var context = new PluginContext(@"c:\scripts\foo.sql", new List<IRuleException>(), textReader);
+
+            var scriptPath = GetTestFilePath(@"c:\scripts\foo.sql");
+            var context = new PluginContext(scriptPath, new List<IRuleException>(), textReader);
 
             var versionWrapper = Substitute.For<IFileversionWrapper>();
             versionWrapper.GetVersion(Arg.Any<Assembly>()).Returns("1.2.3");
@@ -266,7 +262,8 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
         public void ActivatePlugins_ThrowErrors_ShouldCatch_ShouldReport()
         {
             // arrange
-            var path = Path.GetFullPath(Path.Combine(TestContext.CurrentContext.TestDirectory, @"UnitTests/PluginHandler/tsqllint-plugin-throws-exception.dll"));
+            var testFilePath = GetTestFilePath(@"UnitTests/PluginHandler/tsqllint-plugin-throws-exception.dll");
+            var path = Path.GetFullPath(Path.Combine(TestContext.CurrentContext.TestDirectory, testFilePath));
             var assemblyWrapper = new AssemblyWrapper();
 
             var pluginPaths = new Dictionary<string, string>
@@ -289,6 +286,22 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
             // assert
             Assert.AreEqual(1, pluginHandler.Plugins.Count);
             reporter.Received().Report(@"There was a problem with plugin: tsqllint_plugin_throws_exception.PluginThatThrows - something bad happened");
+        }
+
+        private string GetTestFilePath(string path)
+        {
+            if (!isExecutingOnWindows)
+            {
+                if (path.Contains(":"))
+                {
+                    var splitPath = path.Split(":");
+                    path = splitPath[1];
+                }
+
+                path = path.Replace("\\", Path.DirectorySeparatorChar.ToString());
+            }
+
+            return path;
         }
 
         public class TestPlugin2 : TestPlugin
@@ -338,17 +351,13 @@ namespace TSQLLint.Tests.UnitTests.PluginHandler
                     return new[] { defaultPlugin };
                 }
 
-                switch (pathsToPluginNumber[assemblyLoaded])
+                return pathsToPluginNumber[assemblyLoaded] switch
                 {
-                    case 0:
-                        return new[] { defaultPlugin };
-                    case 1:
-                        return new[] { typeof(TestPlugin2) };
-                    case 2:
-                        return new[] { typeof(TestPlugin3) };
-                    default:
-                        return new[] { typeof(TestPlugin4) };
-                }
+                    0 => new[] { defaultPlugin },
+                    1 => new[] { typeof(TestPlugin2) },
+                    2 => new[] { typeof(TestPlugin3) },
+                    _ => new[] { typeof(TestPlugin4) }
+                };
             }
         }
     }
