+++ source/TSQLLint.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27004.2010
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSQLLint.Console", "TSQLLint.Console\TSQLLint.Console.csproj", "{7697E982-E7EA-4D1A-A26C-3DDA32423F1C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSQLLint", "TSQLLint\TSQLLint.csproj", "{7697E982-E7EA-4D1A-A26C-3DDA32423F1C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSQLLint.Tests", "TSQLLint.Tests\TSQLLint.Tests.csproj", "{B3139814-0754-4549-BC6D-9D4F4C772056}"
 EndProject
