+++ source/TSQLLint/appsettings.json
